\name{metaMSdata-package}
\alias{metaMSdata-package}
\alias{metaMSdata}
\docType{package}
\title{Example CDF data for the metaMS package}
\description{
Example CDF data for the metaMS package. For both GC-MS and LC-MS data,
four data files are provided, all in the \code{inst} directory of the
package. In addition, the CDF_GC subdirectory contains a csv file that
can be used as a template for building in-house databases. No further R
code is present.
}
\details{
\tabular{ll}{
Package: \tab metaMSdata\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2013-08-26\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
Built: \tab R 3.0.2; ; 2013-10-09 14:45:48 UTC; unix\cr
}
}
\author{
Pietro Franceschi and Ron Wehrens

Maintainer: Ron Wehrens <ron.wehrens@fmach.it>
}
\keyword{ package }
\seealso{\code{\link[metaMS:metaMS-package]{metaMS}}}
