\name{metaMSdata-package}
\alias{metaMSdata-package}
\alias{metaMSdata}
\docType{package}
\title{Example CDF data for the metaMS package}
\description{
Example CDF data for the metaMS package. For both GC-MS and LC-MS data,
four data files are provided, all in the \code{inst/extdata} directory of the
package. In addition, this directory contains a csv file that
can be used as a template for building in-house databases for GC-MS. No
further R code is present.
}
\author{
Pietro Franceschi and Ron Wehrens

Maintainer: Ron Wehrens <ron.wehrens@fmach.it>
}
\keyword{ package }
\seealso{\code{\link[metaMS:metaMS-package]{metaMS}}}
